GameLogics

Description

Module defines all types and functions needed to store and manipulate gamestate. These functions are used in both Server and MCTS modules to define possible continuation of the current game or state in which the game has ended.

Documentation

data Player

Players : R - Red, Y - Yellow. Red one starts the game.

Constructors

R

Y

type Board = [[Player]]

Current state of the game - every column is described as stack of players' tokens.

data Result

Result of the game.

Constructors

Win Player

Draw

InProgress

data Position

Position is a wrapper for board type, so it is possible to get player whose current turn it is without counting tokens on board.

Constructors

Position

Fields

turn :: Player

board :: Board

newGame :: Position

Creates new game with zero tokens and Red's turn.

makeMove

Arguments

:: Int
number of column to put token in

-> Position
start postion

-> Position
position after move

Adds token to a specified column and changes turn.

legalMoves :: Board -> [Int]

Takes board and returns all numbers of columns where it is possible to put token.

evaluatePosition :: Position -> Result

Takes position and returns whether game has ended and if so, which result it had.

boardToString :: Board -> String

User-friendly view of board.