MCTS

Description

Module represents Haskell implementation of Monte Carlo Tree Search. Implemented MCTS has classic configuration: after adding node to a tree only one rollout is executed, selection is done by chosing any non-visited node, or, if all are visited, by using UCB formula.

Documentation

doMCTS

Arguments

:: StdGen
StdGen used for randomness

-> Position
Game position to evaluate

-> Int
Number of milliseconds to evaluate for

-> IO (Int, StdGen)
(number of column that represents the best found way to put token, StdGen returned from all random actions)

Main function of the module - evaluates position using Monte Carlo Tree Serach for a given number of milliseconds.